Moved the entry point to the program to a separate class. // Перенес точку входа в программу в отдельный класс.
