+++ src/main/java/academy/kata/calculator/Main.java
@@ -5,10 +5,6 @@ import academy.kata.calculator.enums.NumberSystem;
 import academy.kata.calculator.exepcions.CalculatorException;
 import academy.kata.calculator.logics.Calculator;
 import academy.kata.calculator.logics.Validator;
-import academy.kata.calculator.inputProcessors.InConsoleProcessor;
-import academy.kata.calculator.inputProcessors.InProcessor;
-import academy.kata.calculator.outputProcessors.OutConsoleProcessor;
-import academy.kata.calculator.outputProcessors.OutProcessor;
 
 import java.util.NoSuchElementException;
 import java.util.Scanner;
@@ -18,31 +14,9 @@ import static academy.kata.calculator.constants.MainConstants.*;
 
 /**
  * @author Yury Lapitski
- * @version 2024-01-27
+ * @version 2024-01-28
  */
 public class Main {
-
-    public static void main(String[] args) {
-        InProcessor input = new InConsoleProcessor();
-        OutProcessor output = new OutConsoleProcessor();
-
-        while (true) {
-            output.print(INPUT_EXAMPLE_MSG, INPUT_MSG);
-            String inputLine = input.get();
-            output.print(INPUT_EXPRESSION_HEAD_MSG, inputLine, INPUT_EXPRESSION_TAIL_MSG);
-
-            if (inputLine.isEmpty()) {
-                throw new CalculatorException(INPUT_EMPTY_LINE_EXCEPTION_MSG);
-            }
-
-            if (inputLine.equals("q")) {
-                output.print(EXIT_MSG);
-                System.exit(0);   //   completion without error, by code '0'
-            }
-
-            output.print(RESULT_MSG, calc(inputLine), "\n");
-        }
-    }
 
     public static String calc(String input) throws CalculatorException{
         int result;
@@ -68,8 +42,6 @@ public class Main {
                 throw new CalculatorException(INTERNAL_ERROR_EXCEPTION_MSG);
         }
     }
-
-
 
     static NumberSystem getNumberSystem(String expression) throws CalculatorException{
         char ex = expression.trim().charAt(0);
